agents/triage/requirements.txt
confluent-kafka>=2.3.0
openai>=1.12.0
anthropic>=0.18.0
python-dotenv>=1.0.0
boto3>=1.34.0
structlog>=24.1.0
prometheus_client>=0.20.0
